+++ packages/gatsby/index.d.ts
@@ -11,7 +11,6 @@ import {
   ComposeUnionTypeConfig,
 } from "graphql-compose"
 import { GraphQLOutputType } from "graphql"
-import { PluginRef } from "./src/bootstrap/load-plugins/types"
 
 export {
   default as Link,
@@ -47,9 +46,9 @@ export const prefetchPathname: (path: string) => void
  *
  * type IndexQueryProps = { downloadCount: number }
  * type LocaleLookUpInfo = { translationStrings: any } & { langKey: string, slug: string }
- * type IndexPageProps = PageProps<IndexPageProps, LocaleLookUpInfo>
+ * type IndexPageProps = PageProps<IndexQueryProps, LocaleLookUpInfo>
  *
- * export default (props: IndexProps) => {
+ * export default (props: IndexPageProps) => {
  *   ..
  */
 export type PageProps<
@@ -93,9 +92,9 @@ export type PageProps<
    * import {PageProps} from "gatsby"
    *
    * type IndexQueryProps = { downloadCount: number }
-   * type IndexPageProps = PageProps<IndexPageProps>
+   * type IndexPageProps = PageProps<IndexQueryProps>
    *
-   * export default (props: IndexProps) => {
+   * export default (props: IndexPageProps) => {
    *   ..
    *
    */
@@ -110,9 +109,9 @@ export type PageProps<
    *
    * type IndexQueryProps = { downloadCount: number }
    * type LocaleLookUpInfo = { translationStrings: any } & { langKey: string, slug: string }
-   * type IndexPageProps = PageProps<IndexPageProps, LocaleLookUpInfo>
+   * type IndexPageProps = PageProps<IndexQueryProps, LocaleLookUpInfo>
    *
-   * export default (props: IndexProps) => {
+   * export default (props: IndexPageProps) => {
    *   ..
    */
   pageContext: PageContextType
@@ -1538,4 +1537,18 @@ export interface Page<TContext = Record<string, unknown>> {
   matchPath?: string
   component: string
   context: TContext
+}
+
+export interface IPluginRefObject {
+  resolve: string
+  options?: IPluginRefOptions
+  parentDir?: string
+}
+
+export type PluginRef = string | IPluginRefObject
+
+export interface IPluginRefOptions {
+  plugins?: PluginRef[]
+  path?: string
+  [option: string]: unknown
 }