+++ packages/gatsby/src/commands/develop-process.ts
@@ -2,7 +2,6 @@ import url from "url"
 import fs from "fs"
 import openurl from "better-opn"
 import chokidar from "chokidar"
-import { SchemaComposer } from "graphql-compose"
 
 import webpackHotMiddleware from "webpack-hot-middleware"
 import webpackDevMiddleware from "webpack-dev-middleware"
@@ -13,7 +12,7 @@ import webpack from "webpack"
 import graphqlHTTP from "express-graphql"
 import graphqlPlayground from "graphql-playground-middleware-express"
 import graphiqlExplorer from "gatsby-graphiql-explorer"
-import { formatError, GraphQLSchema } from "graphql"
+import { formatError } from "graphql"
 
 import webpackConfig from "../utils/webpack.config"
 import bootstrap from "../bootstrap"
@@ -44,7 +43,6 @@ import sourceNodes from "../utils/source-nodes"
 import { createSchemaCustomization } from "../utils/create-schema-customization"
 import { rebuild as rebuildSchema } from "../schema"
 import { websocketManager } from "../utils/websocket-manager"
-import getSslCert from "../utils/get-ssl-cert"
 import { slash } from "gatsby-core-utils"
 import { initTracer } from "../utils/tracer"
 import apiRunnerNode from "../utils/api-runner-node"
@@ -70,9 +68,6 @@ import {
   runStaticQueries,
   runPageQueries,
 } from "../services"
-
-// checks if a string is a valid ip
-const REGEX_IP = /^(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])\.){3}(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])$/
 
 // const isInteractive = process.stdout.isTTY
 
@@ -204,16 +199,13 @@ async function startServer(program: IProgram): Promise<IServer> {
     graphqlEndpoint,
     graphqlHTTP(
       (): graphqlHTTP.OptionsData => {
-        const {
-          schema,
-          schemaCustomization,
-        }: {
-          schema: GraphQLSchema
-          schemaCustomization: {
-            composer: SchemaComposer<any>
-            context: any
-          }
-        } = store.getState()
+        const { schema, schemaCustomization } = store.getState()
+
+        if (!schemaCustomization.composer) {
+          throw new Error(
+            `A schema composer was not created in time. This is likely a gatsby bug. If you experienced this please create an issue.`
+          )
+        }
 
         return {
           schema,
@@ -359,15 +351,14 @@ async function startServer(program: IProgram): Promise<IServer> {
 
   /**
    * Set up the HTTP server and socket.io.
-   * If a SSL cert exists in program, use it with `createServer`.
    **/
-  const server = program.ssl
-    ? https.createServer(program.ssl, app)
-    : new http.Server(app)
+  const server = new http.Server(app)
 
   const socket = websocketManager.init({ server, directory: program.directory })
 
-  const listener = server.listen(program.port, program.host)
+  // hardcoded `localhost`, because host should match `target` we set
+  // in http proxy in `develop-proxy`
+  const listener = server.listen(program.port, `localhost`)
 
   // Register watcher that rebuilds index.html every time html.js changes.
   const watchGlobs = [`src/html.js`, `plugins/**/gatsby-ssr.js`].map(path =>
@@ -413,14 +404,6 @@ module.exports = async (program: IProgram): Promise<void> => {
   const port =
     typeof program.port === `string` ? parseInt(program.port, 10) : program.port
 
-  // In order to enable custom ssl, --cert-file --key-file and -https flags must all be
-  // used together
-  if ((program[`cert-file`] || program[`key-file`]) && !program.https) {
-    report.panic(
-      `for custom ssl --https, --cert-file, and --key-file must be used together`
-    )
-  }
-
   try {
     program.port = await detectPortInUseAndPrompt(port)
   } catch (e) {
@@ -429,29 +412,6 @@ module.exports = async (program: IProgram): Promise<void> => {
     }
 
     throw e
-  }
-
-  // Check if https is enabled, then create or get SSL cert.
-  // Certs are named 'devcert' and issued to the host.
-  if (program.https) {
-    const sslHost =
-      program.host === `0.0.0.0` || program.host === `::`
-        ? `localhost`
-        : program.host
-
-    if (REGEX_IP.test(sslHost)) {
-      report.panic(
-        `You're trying to generate a ssl certificate for an IP (${sslHost}). Please use a hostname instead.`
-      )
-    }
-
-    program.ssl = await getSslCert({
-      name: sslHost,
-      caFile: program[`ca-file`],
-      certFile: program[`cert-file`],
-      keyFile: program[`key-file`],
-      directory: program.directory,
-    })
   }
 
   // Start bootstrap process.
@@ -678,7 +638,7 @@ module.exports = async (program: IProgram): Promise<void> => {
     // them in a readable focused way.
     const messages = formatWebpackMessages(stats.toJson({}, true))
     const urls = prepareUrls(
-      program.ssl ? `https` : `http`,
+      program.https ? `https` : `http`,
       program.host,
       program.proxyPort
     )