+++ packages/gatsby/src/redux/reducers/index.ts
@@ -22,8 +22,8 @@ import programReducer from "./program"
 import { resolvedNodesCacheReducer } from "./resolved-nodes"
 import { nodesTouchedReducer } from "./nodes-touched"
 import { flattenedPluginsReducer } from "./flattened-plugins"
-import schemaCustomizationReducer from "./schema-customization"
-import inferenceMetadataReducer from "./inference-metadata"
+import { schemaCustomizationReducer } from "./schema-customization"
+import { inferenceMetadataReducer } from "./inference-metadata"
 
 /**
  * @property exports.nodesTouched Set<string>